Move javascript messages to enum and add tests
